backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Seba",email = "sebsym96@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=6.0.0,<7.0.0)",
    "wagtail (>=7.2.1,<8.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "stripe (>=11.0.0,<12.0.0)",
    "djangorestframework (>=3.15.0,<4.0.0)",
]

[dependency-groups]
dev = [
    "black (>=25.12.0,<26.0.0)",
    "pylint (>=4.0.4,<5.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
